requirements.txt
ijson
pickle
newspaper3k